requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.2
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
